Fix sudoders permissions, add upstart job, systemctl call if present

+++ packaging/scripts/preremove.sh
@@ -9,16 +9,18 @@ esac
 case "$1" in
     upgrade)
         # On Debian, we stop in pre-remove.
-        systemctl stop glouton.service
+	test -e /lib/init/upstart-job && stop glouton
+        test -x /usr/bin/systemctl -o -x /bin/systemctl && systemctl stop glouton.service
         ;;
     remove)
-        systemctl --no-reload disable glouton.service
-        systemctl stop glouton.service
+	test -e /lib/init/upstart-job && stop glouton
+        test -x /usr/bin/systemctl -o -x /bin/systemctl && systemctl --no-reload disable glouton.service
+        test -x /usr/bin/systemctl -o -x /bin/systemctl && systemctl stop glouton.service
         ;;
     0)
         # On CentOS, we only stop if uninstall in pre-remove and restart in post-remove
-        systemctl --no-reload disable glouton.service > /dev/null 2>&1 || :
-        systemctl stop glouton.service > /dev/null 2>&1 || :
+        test -x /usr/bin/systemctl -o -x /bin/systemctl && systemctl --no-reload disable glouton.service > /dev/null 2>&1 || :
+        test -x /usr/bin/systemctl -o -x /bin/systemctl && systemctl stop glouton.service > /dev/null 2>&1 || :
         ;;
 esac
